MH3-24650: Regenerated SAM E70 Legato Adventure for gfx-v3.6.0 release

+++ apps/legato_adventure/firmware/src/config/ssd1963_rgb565_mxt_e70/gfx/legato/common/legato_common.h
@@ -323,10 +323,10 @@ static const leSize leSize_Zero = {0, 0};
 */
 typedef struct leRect
 {
-    int32_t x;
-    int32_t y;
-    int32_t width;
-    int32_t height;
+    int16_t x;
+    int16_t y;
+    int16_t width;
+    int16_t height;
 } leRect;
 
 typedef void* leBuffer;
